requirements.txt
pandas
requests
clickhouse-connect
apache-airflow